Move action helpers to actions module
